+++ tests/utils/test_describe_model.py
@@ -102,6 +102,7 @@ class TestDescribeModel(test.SimpleTestCase):
                 "description": None,
                 "docstring": None,
                 "constraints": {},
+                "db_constraint": True,
                 "raw_field": None,
             },
         )
@@ -124,6 +125,7 @@ class TestDescribeModel(test.SimpleTestCase):
                 "docstring": None,
                 "constraints": {},
                 "raw_field": None,
+                "db_constraint": True,
             },
         )
 
@@ -251,6 +253,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "default": None,
                         "description": "Tree!",
                         "docstring": None,
+                        "db_constraint": True,
                         "constraints": {},
                     }
                 ],
@@ -267,6 +270,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": "Tree!",
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "o2o_fields": [
@@ -284,6 +288,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "raw_field": "o2o_id",
                         "unique": True,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "backward_o2o_fields": [
@@ -299,6 +304,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "python_type": "models.StraightFields",
                         "unique": False,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "m2m_fields": [
@@ -321,6 +327,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "through": "straightfields_straightfields",
                         "on_delete": "CASCADE",
                         "_generated": False,
+                        "db_constraint": True,
                     },
                     {
                         "name": "rel_from",
@@ -340,6 +347,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "backward_key": "straightfields_id",
                         "through": "straightfields_straightfields",
                         "on_delete": "CASCADE",
+                        "db_constraint": True,
                         "_generated": True,
                     },
                 ],
@@ -465,6 +473,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": "Tree!",
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "backward_fk_fields": [
@@ -480,6 +489,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": "Tree!",
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "o2o_fields": [
@@ -497,6 +507,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "raw_field": "o2o_id",
                         "unique": True,
                         "constraints": {},
+                        "db_constraint": True,
                     },
                 ],
                 "backward_o2o_fields": [
@@ -512,6 +523,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "python_type": StraightFields,
                         "unique": False,
                         "constraints": {},
+                        "db_constraint": True,
                     },
                 ],
                 "m2m_fields": [
@@ -534,6 +546,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "through": "straightfields_straightfields",
                         "on_delete": "CASCADE",
                         "_generated": False,
+                        "db_constraint": True,
                     },
                     {
                         "name": "rel_from",
@@ -554,6 +567,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "through": "straightfields_straightfields",
                         "on_delete": "CASCADE",
                         "_generated": True,
+                        "db_constraint": True,
                     },
                 ],
             },
@@ -678,6 +692,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": "Tree!",
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "backward_fk_fields": [
@@ -693,6 +708,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": "Tree!",
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "o2o_fields": [
@@ -710,6 +726,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "raw_field": "o2o_id",
                         "unique": True,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "backward_o2o_fields": [
@@ -725,6 +742,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "python_type": "models.SourceFields",
                         "unique": False,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "m2m_fields": [
@@ -747,6 +765,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "through": "sometable_self",
                         "on_delete": "CASCADE",
                         "_generated": False,
+                        "db_constraint": True,
                     },
                     {
                         "name": "rel_from",
@@ -767,6 +786,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "through": "sometable_self",
                         "on_delete": "CASCADE",
                         "_generated": True,
+                        "db_constraint": True,
                     },
                 ],
             },
@@ -891,6 +911,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": "Tree!",
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "backward_fk_fields": [
@@ -906,6 +927,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": "Tree!",
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "o2o_fields": [
@@ -923,6 +945,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "raw_field": "o2o_id",
                         "unique": True,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "backward_o2o_fields": [
@@ -938,6 +961,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "python_type": SourceFields,
                         "unique": False,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "m2m_fields": [
@@ -960,6 +984,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "through": "sometable_self",
                         "on_delete": "CASCADE",
                         "_generated": False,
+                        "db_constraint": True,
                     },
                     {
                         "name": "rel_from",
@@ -980,6 +1005,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "through": "sometable_self",
                         "on_delete": "CASCADE",
                         "_generated": True,
+                        "db_constraint": True,
                     },
                 ],
             },
@@ -1028,6 +1054,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": None,
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     },
                 ],
                 "o2o_fields": [],
@@ -1037,6 +1064,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "_generated": True,
                         "backward_key": "uuidpkmodel_id",
                         "constraints": {},
+                        "db_constraint": True,
                         "default": None,
                         "description": None,
                         "docstring": None,
@@ -1099,6 +1127,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "description": None,
                         "docstring": None,
                         "constraints": {},
+                        "db_constraint": True,
                     },
                 ],
                 "o2o_fields": [],
@@ -1106,6 +1135,7 @@ class TestDescribeModel(test.SimpleTestCase):
                 "m2m_fields": [
                     {
                         "name": "peers",
+                        "db_constraint": True,
                         "generated": False,
                         "nullable": False,
                         "field_type": ManyToManyFieldInstance,
@@ -1202,6 +1232,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "raw_field": "model_id",
                         "unique": False,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "m2m_fields": [],
@@ -1221,6 +1252,7 @@ class TestDescribeModel(test.SimpleTestCase):
                         "raw_field": "parent_id",
                         "unique": True,
                         "constraints": {},
+                        "db_constraint": True,
                     }
                 ],
                 "pk_field": {