+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
 - Add `RawSQL` expression.
 - Fix columns count with annotations in `_make_query`. (#776)
 - Make functions nested. (#828)
+- Add `db_constraint` in field describe.
 
 0.17.5
 ------